move rlist command to archiver.rlist
