+++ setup.cfg
@@ -127,6 +127,7 @@ per_file_ignores =
     src/borg/archiver/keys.py:F405
     src/borg/archiver/prune.py:F405
     src/borg/archiver/rdelete.py:F405
+    src/borg/archiver/rlist.py:E501
     src/borg/archiver/tar.py:F405
     src/borg/cache.py:E127,E128,E402,E501,E722,W504
     src/borg/fuse.py:E402,E501,E722,W504
